auth_service/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
python-jose
pydantic
passlib[bcrypt]